+++ layrz/sdk/entities/reports/__init__.py
@@ -2,6 +2,6 @@
 from .col import ReportCol, ReportDataType
 from .format import ReportFormat
 from .header import ReportHeader
-from .page import ReportPage
+from .page import ReportPage, CustomReportPage
 from .report import Report
 from .row import ReportRow